+++ hikka/modules/hikka_backup.py
@@ -19,7 +19,7 @@ from pathlib import Path
 
 import redis
 
-from hikkatl.tl.types import Message
+from telethon.tl.types import Message
 
 from .. import loader, main, utils
 from ..inline.types import BotInlineCall